Use float literals instead of casting

+++ heap_test.go
@@ -9,15 +9,15 @@ import (
 
 func TestHeap(t *testing.T) {
 	heap := newHeap([]float64{1, 2, 3, 4}, true)
-	assert.Equal(t, float64(4), heapops.Pop(heap))
+	assert.Equal(t, 4., heapops.Pop(heap))
 
 	heap = newHeap([]float64{1, 2, 3, 4}, false)
-	assert.Equal(t, float64(1), heapops.Pop(heap))
+	assert.Equal(t, 1., heapops.Pop(heap))
 }
 
 func TestPeek(t *testing.T) {
 	heap := newHeap([]float64{1, 2, 3, 4}, true)
-	heapops.Push(heap, float64(5))
-	heapops.Push(heap, float64(4))
-	assert.Equal(t, float64(5), heap.peek())
+	heapops.Push(heap, 5.)
+	heapops.Push(heap, 4.)
+	assert.Equal(t, 5., heap.peek())
 }
